# pkg/gen/http/openapi.json
{"swagger":"2.0","info":{"title":"Books Service","description":"HTTP service for working with books","version":""},"host":"localhost:8088","consumes":["application/json","application/xml","application/gob"],"produces":["application/json","application/xml","application/gob"],"paths":{"/books/list":{"get":{"tags":["books"],"summary":"list books","description":"List books","operationId":"books#list","responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/BooksListResponseBody"}}},"schemes":["http"]}},"/books/pickup/{book_id}":{"post":{"tags":["books"],"summary":"pickup books","description":"Mark book as picked up\n\n**Required security scopes for jwt**:\n  * `books:pickup`","operationId":"books#pickup","parameters":[{"name":"book_id","in":"path","description":"id of the Book","required":true,"type":"integer","format":"int64"},{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"},{"name":"PickupRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/BooksPickupRequestBody","required":["subscriber_id"]}}],"responses":{"200":{"description":"OK response.","schema":{}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]}},"/books/reserve/{book_id}":{"post":{"tags":["books"],"summary":"reserve books","description":"Mark book as reserved. Once a book is reserved timer starts with timeout for the book to become picked up. Timeout is configurable. \n\t\tOnce timeout is expired book becomes available\n\n**Required security scopes for jwt**:\n  * `books:reserve`","operationId":"books#reserve","parameters":[{"name":"book_id","in":"path","description":"id of the Book","required":true,"type":"integer","format":"int64"},{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"},{"name":"ReserveRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/BooksReserveRequestBody","required":["subscriber_id"]}}],"responses":{"200":{"description":"OK response.","schema":{}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]}},"/books/return":{"post":{"tags":["books"],"summary":"return books","description":"Mark book as returned\n\n**Required security scopes for jwt**:\n  * `books:return`","operationId":"books#return","parameters":[{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"},{"name":"ReturnRequestBody","in":"body","required":true,"schema":{"$ref":"#/definitions/BooksReturnRequestBody","required":["book_id","subscriber_id"]}}],"responses":{"200":{"description":"OK response.","schema":{}},"403":{"description":"Forbidden response.","schema":{"$ref":"#/definitions/BooksReturnInvalidScopesResponseBody"}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]}},"/books/subscribe/{book_id}":{"post":{"tags":["books"],"summary":"subscribe books","description":"Subscribe the caller on the next 'book's become available\n\n**Required security scopes for jwt**:\n  * `books:subscribe`","operationId":"books#subscribe","parameters":[{"name":"book_id","in":"path","description":"id of the Book","required":true,"type":"integer","format":"int64"},{"name":"Authorization","in":"header","description":"JWT used for authentication","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{}},"403":{"description":"Forbidden response.","schema":{"$ref":"#/definitions/BooksSubscribeInvalidScopesResponseBody"}}},"schemes":["http"],"security":[{"jwt_header_Authorization":[]}]}},"/signin":{"post":{"tags":["auther"],"summary":"signin auther","description":"Creates a valid JWT","operationId":"auther#signin","parameters":[{"name":"Authorization","in":"header","description":"Basic Auth security using Basic scheme (https://tools.ietf.org/html/rfc7617)","required":true,"type":"string"}],"responses":{"200":{"description":"OK response.","schema":{"$ref":"#/definitions/AutherSigninResponseBody","required":["jwt"]}},"401":{"description":"Unauthorized response.","schema":{"$ref":"#/definitions/AutherSigninUnauthorizedResponseBody"}}},"schemes":["http"],"security":[{"basic_header_Authorization":[]}]}}},"definitions":{"AutherSigninResponseBody":{"title":"AutherSigninResponseBody","type":"object","properties":{"jwt":{"type":"string","description":"JWT token","example":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"}},"example":{"jwt":"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ"},"required":["jwt"]},"AutherSigninUnauthorizedResponseBody":{"title":"AutherSigninUnauthorizedResponseBody","type":"string","description":"Credentials are invalid","example":"Accusamus ut corrupti voluptatem voluptatem."},"BookResponseBody":{"title":"BookResponseBody","type":"object","properties":{"annotation":{"type":"string","example":"Est aut minus et enim odio."},"author":{"type":"string","example":"Ut consequatur quia consequatur et sed."},"id":{"type":"integer","example":6639163768559936509,"format":"int64"},"images":{"type":"array","items":{"type":"string","example":"Doloribus iste accusantium quo aliquam sit."},"description":"images are a list of book photos","example":["Consequatur fuga voluptatibus alias et fugit aut.","Quas et.","Veritatis delectus voluptatem tenetur eveniet earum voluptatem."]},"status":{"type":"string","example":"Error officiis et."},"title":{"type":"string","example":"Et quasi et qui et deleniti."}},"description":"Book is a book","example":{"annotation":"Repellendus rerum alias distinctio.","author":"Occaecati sequi veniam.","id":3754138419164672964,"images":["Voluptate velit ratione sint.","Quos praesentium sed nihil fugit minima.","Consequatur non cupiditate consectetur architecto aut."],"status":"Quis distinctio labore optio incidunt.","title":"Quas itaque sunt."},"required":["id","title","annotation","author","images","status"]},"BooksListResponseBody":{"title":"Mediatype identifier: application/library.books; view=default","type":"object","properties":{"data":{"type":"array","items":{"$ref":"#/definitions/BookResponseBody"},"example":[{"annotation":"Magni et veniam est quibusdam rerum.","author":"Dolor labore.","id":6236022153690076209,"images":["Exercitationem exercitationem.","Harum eaque cum quia eos repellendus itaque."],"status":"Quam rerum enim quibusdam.","title":"Saepe sunt est dolor accusantium expedita commodi."},{"annotation":"Magni et veniam est quibusdam rerum.","author":"Dolor labore.","id":6236022153690076209,"images":["Exercitationem exercitationem.","Harum eaque cum quia eos repellendus itaque."],"status":"Quam rerum enim quibusdam.","title":"Saepe sunt est dolor accusantium expedita commodi."}]}},"description":"ListResponseBody result type (default view)","example":{"data":[{"annotation":"Magni et veniam est quibusdam rerum.","author":"Dolor labore.","id":6236022153690076209,"images":["Exercitationem exercitationem.","Harum eaque cum quia eos repellendus itaque."],"status":"Quam rerum enim quibusdam.","title":"Saepe sunt est dolor accusantium expedita commodi."},{"annotation":"Magni et veniam est quibusdam rerum.","author":"Dolor labore.","id":6236022153690076209,"images":["Exercitationem exercitationem.","Harum eaque cum quia eos repellendus itaque."],"status":"Quam rerum enim quibusdam.","title":"Saepe sunt est dolor accusantium expedita commodi."},{"annotation":"Magni et veniam est quibusdam rerum.","author":"Dolor labore.","id":6236022153690076209,"images":["Exercitationem exercitationem.","Harum eaque cum quia eos repellendus itaque."],"status":"Quam rerum enim quibusdam.","title":"Saepe sunt est dolor accusantium expedita commodi."}]},"required":["data"]},"BooksPickupRequestBody":{"title":"BooksPickupRequestBody","type":"object","properties":{"subscriber_id":{"type":"string","description":"id of a subscriber picking up the book","example":"Sit corrupti saepe laudantium qui eligendi."}},"example":{"subscriber_id":"Fuga est sed blanditiis et."},"required":["subscriber_id"]},"BooksReserveRequestBody":{"title":"BooksReserveRequestBody","type":"object","properties":{"subscriber_id":{"type":"string","description":"id of a subscriber picking up the book","example":"Non id rem libero a sunt."}},"example":{"subscriber_id":"Qui aut doloribus expedita autem quis iusto."},"required":["subscriber_id"]},"BooksReturnInvalidScopesResponseBody":{"title":"BooksReturnInvalidScopesResponseBody","type":"string","description":"Token scopes are invalid","example":"Neque voluptas sequi soluta consequatur."},"BooksReturnRequestBody":{"title":"BooksReturnRequestBody","type":"object","properties":{"book_id":{"type":"integer","description":"id of the Book","example":1588809184003176921,"format":"int64"},"subscriber_id":{"type":"string","description":"id of a subscriber returning the book","example":"Ut quis hic et dicta et eum."}},"example":{"book_id":7548870960549249963,"subscriber_id":"Exercitationem ex odit."},"required":["book_id","subscriber_id"]},"BooksSubscribeInvalidScopesResponseBody":{"title":"BooksSubscribeInvalidScopesResponseBody","type":"string","description":"Token scopes are invalid","example":"Est ut est sit et perferendis."}},"securityDefinitions":{"basic_header_Authorization":{"type":"basic","description":"Basic authentication used to authenticate security principal during signin"},"jwt_header_Authorization":{"type":"apiKey","description":"Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint. \n\tSupports scopes \"books:reserve\", \"books:pickup\", \"books:return\", \"books:subscribe\"\n\n**Security Scopes**:\n  * `books:reserve`: reserve access\n  * `books:pickup`: pickup access\n  * `books:return`: return access\n  * `books:subscribe`: subscribe access","name":"Authorization","in":"header"}}}